eupv: Don't error out on missing flatpaks
In a perfect world, USB drives created with eos-updater-prepare-volume
could be used to update Endless computers on any previous OS versions,
but alas, this is not a perfect world. One of the reasons we can't keep
that promise is checkpoints ("eos.checkpoint-target"), but we also can't
guarantee we'll have all the flatpaks needed by
eos-updater-flatpak-installer on the target machine receiving the
update. This is because flatpaks may be uninstalled by the user after
being auto-installed, or they may be installed and later uninstalled by
eos-updater-flatpak-installer, as will be the case for
com.hack_computer.ProjectLibrary in 3.9.1.

So make failures of the `flatpak create-usb` command non-fatal for
flatpaks not explicitly specified by the user. We may also want to
install any missing flatpaks before doing the copy when we have an
Internet connection, but that improvement is left as an exercise for the
reader.

https://phabricator.endlessm.com/T30368