eos-updater-flatpak-installer: Add a fallback service and timer
This service will be run once a day during normal system use (not just
during the first boot after an upgrade, as e-u-f-i is normally run),
just in case there were errors in downloading or installing the updates
during the most recent system upgrade.

The fallback runs e-u-f-i with `--pull`, so it will download flatpaks,
runtimes, etc. if necessary. For this reason it needs to be slightly
less constrained in its systemd sandbox than
`eos-updater-flatpak-installer.service`. Other than that the service
files are almost identical.

This could be considered a hack, but could also be considered as a
safety fallback in case of mistakes in previous OS versions. Otherwise,
the e-u-f-i process is only run with the eos-updater/OSTree/flatpak code
from release X just after upgrading to release X, or just before
upgrading to release X+Y. If there are problems in any of that code, the
user has to do a successful OS upgrade to get a fixed version of the
code, and it’s a month (or more) between OS releases.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T31190