test-common: Use standalone ostree-trivial-httpd
The `trivial-httpd` builtin for the `ostree` command has been deprecated
for a couple years and is not included on default builds of ostree. On
the other hand, the standalone `ostree-trivial-httpd` has been available
since ostree 2017.1 in `<libexecdir>/libostree`. It may later move to
`<libexecdir>/installed-tests/libostree`. Try to use those paths instead
of the builtin.

https://phabricator.endlessm.com/T31384