eos-updater: Don’t follow checkpoints if running an old system
This puts the infrastructure in place for not following a checkpoint
redirect if running on an old system which is not supported by the
target ref.

The following commits will add some details of systems to ignore.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T31918
(cherry picked from commit fbf4d9dc68e2c8f69bd903d1edf65a5cdda6737b)