tests: Remove indexed summary for invalid collection ID test
With newer flatpak, an indexed summary file will be created on the
server and the client will try to download `summary.idx` first. This
breaks the test that expects to fail when there's no collection ID in
the remote summary since flatpak only validates its presence on a full
`summary` file.

Since the Endless servers aren't providing indexed summaries yet, it's
useful to continue simulating this test with a full summary that doesn't
contain a collection ID. If the summary index exists, delete it to go
back to older flatpak behavior.

https://phabricator.endlessm.com/T31918