eos-updater: Tweak eos3a → eos4 checkpoint tests on aarch64 hosts
Since the checkpoint redirect should not be followed on aarch64 hosts,
the tests are going to behave slightly differently on that architecture:
we should expect the checkpoint to never be followed unless the
architecture is mocked, or following redirects is forced.

The alternative approach would be to always mock the architecture, and
set it to something other than aarch64 (unless explicitly testing the
architecture checks), but then we’d be throwing away an opportunity to
test the checkpoint code on exactly the platform it’s meant to catch.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T31918
(cherry picked from commit 17e9b623bac9af8fd34eb593cc238a89327ea9c4)