eos-updater: Fix for upgrading to eos4 and newer's ref
Users may upgrading system from booted ref eos3a to both eos4, or even
newer target ref in the future. The newer target ref is defined as
"latest1" in T32550.
The target ref is always newer than booted ref. So, it is fine to only
check that the booted ref is eos3a.

Also fixed the updating condition test plans.

https://phabricator.endlessm.com/T32552