flatpak-installer: Add prune-ref action for dangling OSTree refs
There have been some bugs where a Flatpak is uninstalled but the
underlying OSTree ref is not removed. That means the commit objects
won't be pruned and the host won't recover any disk space. This adds the
`prune-ref` autoinstall action to remove those refs. This action only
takes effect when the corresponding Flatpak is no longer installed.

https://phabricator.endlessm.com/T32791