test-common: Ensure client receives updated summary
To try to avoid downloading the summary file when it already has the
current version, the ostree client requests the summary and signature
with an If-Modified-Since HTTP header (when it can't use the preferable
If-None-Match header). The HTTP header only has second precision, so it
may not receive an updated summary if the request is sent within the
same second the summary is updated.

To ensure the client will always receive the updated summary, sleep
until the next second if necessary. Note that only the signature file is
checked since it's always created after the summary.