libeos-updater-util: Don't leak ostree checkout tmpdir
Only a subdirectory of the temporary directory was being passed back to
the caller, so nothing was cleaning it up. Instead, have the caller
allocate the temporary directory and provide the path to perform the
checkout in. Without this, `/tmp/ostree-checkout-*` directories were
being left behind every time `eos-updater` entered the fetch state.