eos-updater: Add support for OSTree sysroot autocleanup
When OSTree staged deployments are used, the old rollback deployment is
deleted during system shutdown. To keep from slowing down shutdown, the
OSTree repo is not pruned at that time. That means that even though the
deployment was deleted, the objects are still on disk. Since that may be
a significant amount of wasted disk space, the full cleanup with repo
pruning needs to be run at some time after rebooting. See
https://github.com/ostreedev/ostree/issues/2510 for details.

To detect when cleanup is necessary, a systemd drop in is added to touch
the `/sysroot/.cleanup` file after `ostree-finalize-staged.service` has
finalized the new deployment. The reason to use a drop-in for
`ostree-finalize-staged.service` rather then creating the file from
`eos-updater` is to avoid the situation where an unclean shutdown occurs
and the new deployment is not finalized. In that case, cleanup would be
run unnecessarily on the next boot.

A new systemd service, `eos-updater-autocleanup.service`, is added to
run `ostree admin cleanup` when `/sysroot/.cleanup` exists and then
delete it afterwards. This adds a dependency on the `ostree` CLI but a
separate program could be provided calling the `ostree_sysroot_cleanup`
API and deleting the `/sysroot/.cleanup` file itself.

https://phabricator.endlessm.com/T5658