libeos-updater-util: Add method to detect if /boot is an automount
On some hosts, systemd sets up `/boot` as an autofs automount.
Unfortunately, this configuration has some bugs that make usage of
`/boot` less than robust.

Add a method to see if `/boot` is an automount. This uses libmount to
parse `/proc/self/mountinfo` and check if the first mountpoint for the
sysroot's boot directory has the `autofs` filesystem type. libmount
version 2.24 was chosen as a minimum because that's when `libmnt_table`
and `libmnt_cache` reference counting was added.

https://phabricator.endlessm.com/T33136