tests: Port from g_spawn_check_exit_status() to g_spawn_check_wait_status()
The latter replaced the former in GLib 2.70 to fix the incorrect
nomenclature. An exit status is what you’d return from `main()`, where
zero is success and any non-zero number is failure.

A wait status is what’s returned by `waitpid()` and similar, and is not
the same. It must be decoded using `WIFEXITED()`, `WIFSIGNALED()`, etc.

This bumps our GLib dependency to 2.70, but that’s what’s available on
EOS master already.

This introduces no functional changes, but fixes some compiler warnings.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>