licensing: Add SPDX-License-Identifier to all files
This adds the licensing identifier to all files which contain a license
header already or, for the ones which don’t, add the default license to
them.

The only files this doesn’t modify are those where it’s not
straightforward to add a licensing header due to the file format
(they’ll be handled in a later commit) and the man files (where we
probably don’t want to use the LGPL-2.1+).

This commit does not change the effective license of any of these files,
it only encodes their current licensing.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>