licensing: Add copyright headers to files which were missing them
The copyrights have been manually extracted from the git logs for each
file.

The copyright owner ‘Endless Mobile, Inc.’ was used for dates pre-2020
and the copyright owner ‘Endless OS Foundation, LLC’ for dates since
then, for historical correctness. They are effectively the same as the
copyrights from the former all transferred to the latter when the name
changed.

This commit doesn’t change the copyright (or licensing) status of any
files, it just encodes what’s already there.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>