tests: Really make CheckpointTestData field nullable
If the fields are allowed to be `NULL`, then all accesses of them need
to be prepared for that. The easiest way to do that is to move the
default fallbacks back into the common `do_update_refspec_checkpoint`
function. Only in the case of the eos4 → eos5 test do the refs need to
be handled at a higher level. The vendor and product can certainly be
handled in the common code.

https://phabricator.endlessm.com/T34110