tests: Use named initializers with CheckpointTestData
This struct is getting pretty large and most of the time only one or two
fields are set. Using named initializers makes it much easier to see
what the particular test case is for. Named initializers are a C99
feature and this project uses GNU11, so this should be fine for any
compiler.

https://phabricator.endlessm.com/T34110