tests: Allow non-string metadata in OSTree commits in the unit tests
Various of the unit tests construct temporary OSTrees to test against.
Some of these OSTrees have commits with additional metadata set (such as
end of life rebase metadata or checkpoints). Currently, that metadata is
always a string/string key/value pair.

For some tests we’ll want to add soon, the value has to be a uint64.

This commit changes the test infrastructure and existing tests to use
string/variant key/value pairs. All the variants are strings for now,
but the new test will use a variant with a different type.

This introduces no functional changes into the tests for now.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T34079