eos-updater: Set UserVisible update flag implicitly on checkpoints
While this isn’t a long-term solution (see the FIXME comments), it’ll do
for now for the EOS 4 → 5 transition. It implicitly sets the
`EU_UPDATE_FLAGS_USER_VISIBLE` flag on the commit pointed to by a
checkpoint, without having to change the build and deployment system to
set the flag explicitly.

Includes a unit test.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>
https://phabricator.endlessm.com/T34079