eos-updater: Add a new UpdateIsUserVisible D-Bus property
This will expose whether the commit which is due to be fetched and
deployed contains user visible changes which should be notified to the
user before the update proceeds.

The policy is based on explicit semantics of version numbering: the
major versions of the currently booted commit and candidate update
commit are compared and, iff different, the update contains user visible
changes.

Includes a unit test.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T34079
(cherry picked from commit 403fdb7cbb7fa930249b85df97f5d9d04fe4237f)