eos-updater: Add a new UpdateFlags D-Bus property
This will expose flags for the commit which is due to be fetched and
deployed if the upgrade proceeds.

There’s currently only one flag – whether the commit contains
user-visible changes relative to its parent commit – but more could be
added in future, such as that it contains security updates.

Includes a unit test.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T34079