tests: Set sysroot bootloader to none by default
Until now all the ostree deployment tests have been setting up a fake
u-boot configuration so that ostree's automatic bootloader detection
finds a valid configuration.

However, since ostree 2019.2, it has been possible to specify the
bootloader using the `sysroot.bootloader` repo config key. When the
value is `none`, ostree only updates the boot loader spec entries in
`boot/loader/entries`. When the value is `auto` or unset, it uses the
previous behavior of probing for a bootloader and trying to update the
bootloader's native configuration as well.

Since eos-updater doesn't generally need to be in the business of
validating ostree's bootloader detection, make the tests default to
setting `sysroot.bootloader` to `none`. Automatic detection is kept for
a future test.

https://phabricator.endlessm.com/T34110