tests: Rework D-Bus name watching utility code to fix a race
Previously, the code ran a main loop until a callback within the main
loop quit it. That was fine as long as the callback was executed after
the main loop had been started.

However, in the case that the D-Bus name we care about is already in the
state we’re looking for (typically, not owned), `g_main_loop_quit()`
would be called immediately, before `g_main_loop_run()`, which would
mean the main loop runs forever.

Avoid that, and simplify the logic, by using `GMainContext` directly and
moving the exit condition to the top-level utility function rather than
having it in the callback.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T34079