eos-updater: Don’t propagate update information if update isn’t an update
The return value of `is_checksum_an_update()` doesn’t indicate whether
the checksum is an update, it indicates whether there was an error in
checking whether checksum is an update.

The value of `out_commit` gives whether the checksum is an update.

Avoid returning a nearly-valid `UpdateRefInfo` if the update shouldn’t
actually be used.

This changes this call site of `is_checksum_an_update()` to match the
other existing call site.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T34079