build: Check for flake8 module when finding the Python installation
Since we use it in the unit tests.

It’s required for building the Debian packaging due to being listed in
`Build-Depends`, but might not be pulled in particularly early in the
build if building directly with Meson. Checking for it here will force
failure earlier in the build process.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T26011