eos-updater: Add a ReleaseNotesUri property
This adds a `ReleaseNotesUri` property to the D-Bus interface. This is
populated from an optional `eos-updater.release-notes-uri` metadata key
in OSTree commits, which has a couple of optional string substitutions
applied to it to allow a single URI template to be refined for multiple
possible update paths.

This allows different release notes pages to be shown to users depending
on which version they are updating from.

Includes a unit test.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T34082