eos-autoupdater: Cache poll results on disk
In order to tell when the poll results have changed from a previous run,
store them on disk in `/var/lib/eos-updater/autoupdater-poll-results`.
The semantics of the file are that it stores the last time the poll
results changed. If the update ref and commit don't change on a
subsequent run, then the stored file and last changed time don't change.
This will be used in a subsequent commit to determine whether to fetch
and apply the update or not.

https://phabricator.endlessm.com/T34083