eos-autoupdater: Delay user visible updates
For user visible updates it's beneficial for the user to review the
update in the software center rather than being immediately
automatically applied. To provide this window, add the configuration
option `UserVisibleUpdateDelayDays`. If the update is user visible and
it has been less than that number of days since the update has been
seen, it is not applied. By default this is 14 days.

https://phabricator.endlessm.com/T34083