eos-updater: Allow quitting from all quiescent states
The only reason to delay quitting when the quit file has been deleted is
if the updater is in the middle of one of the active states. Without
this, the updater can't be stopped using the quit file after polling or
fetching when an update is available.

https://phabricator.endlessm.com/T34083
(cherry picked from commit 4a9ffc1ad46e418324821d00e15e7454d5aaf02e)