test-common: Always force autoupdater fetching
When mogwai is installed, eos-updater needs to avoid download scheduling
during testing or it may not fetch the commit. Normally this is done by
running the autoupdater with `--force-update`, but that option enables
several autoupdater policies. When testing the autoupdater that can make
it impossible to discern its behavior.

Introduce a new `--force-fetch` option and always set it when running
the autoupdater from the test suite. Testing of the eos-updater download
scheduling behavior can be done outside the use of the autoupdater if
desired.

https://phabricator.endlessm.com/T34083
(cherry picked from commit c83a55241774edac0236e7bf3ac53f212bbfaa95)