test-common: Convert eos_test_add to function
`eos_test_add` is a simple wrapper around `g_test_add`, which is
implemented as a macro in a statement block. When `eos_test_add` is also
a macro, any literal arguments are constructed within a block that ends
when the call returns. However, `g_test_add` stores the arguments in a
table for later usage when `g_test_run` is called. This prevents use of
compound literals being passed by address since the address is no longer
valid once `g_test_add` returns.

To workaround this, convert `eos_test_add` to a function. That will
ensure any compound literals used as arguments are defined in the
caller's scope. As long as `eos_test_add` and `g_test_run` are called
within the same block, the address of the unnamed object will remain
valid.

https://phabricator.endlessm.com/T34083