eos-autoupdater: Ensure poll results are always read from disk
Previously the stored poll results were only being read from disk after
completing polling. However, if eos-updater was already past the polling
state, then nothing would read the stored results and store them in
memory. Upon attempting to apply a user visible update, it would
segfault dereferencing a `NULL` pointer to the in memory poll results.

If an update is available or ready, update the poll results if they're
not set in memory yet. This should ensure the autoupdater works
correctly if it's not the component causing the updater to poll or if
there are issues reading the stored results.

https://phabricator.endlessm.com/T34083