test-common: Use soup HTTP server
Replace the usage of the `ostree-trivial-httpd` HTTP server with the new
`SoupServer` implementation. The `SoupServer` implementation is simpler
and more efficient to manage since it runs in process, the server data
doesn't need to be scraped from the filesystem, and the server is
explicitly stopped when the structure is freed.