Port to Soup 3
Port the 2 `SoupServer` instances to use Soup 3 instead of Soup 2.
Version 3.2 is specified to use the non-deprecated
`soup_server_message_pause` and `soup_server_message_unpause` methods.
The changes needed are mostly described in the following:

https://libsoup.org/libsoup-3.0/migrating-from-libsoup-2.html
https://libsoup.org/libsoup-3.0/server-howto.html

This is done in one mega-commit to avoid a likely untestable situation
where part of eos-updater is linked to libsoup2 and the other is linked
to libsoup3.

https://phabricator.endlessm.com/T33677