Consolidate meson dependencies at top level
It's kind of nice that each directory states its specific dependency
version requirements, but ultimately it makes hunting down the
requirements hard. It also slows down meson as it repeatedly checks the
same dependency.

Since all the code is going to be built with the same installed
dependencies, just consolidate all of them at the top level.