Split detecting flake8 from detecting python3 installation
In b2ba34fefe0dc72a31efba109222b996da0ee99b I made flake8 optional,
primarily for the benefit of GNOME OS.

However, I overlooked the fact that the 'py3' object is not only used to
run flake8, but also to run some Python tests. As a result those tests
would silently not run if flake8 is not available: not the desired
result.

Instead, search for a 'flake8' executable, separately to finding a
working Python 3 installation for those tests. Make the Python 3
installation mandatory, and flake8 optional. Adjust the Debian
build-dependencies to install the 'flake8' entrypoint.