Run flake8 from source directory
By default, Meson runs tests from within the build directory.

When searching for its configuration file (setup.cfg or tox.ini or
.flake8) you might hope that flake8 would walk the directory hierarchy
upwards from the directory containing each source file it is scanning,
but it actually starts from the current working directory.

When running in jenkins or when building the .deb package, this is OK
because the build directory is a subdirectory of the build directory.

However, when developing with GNOME Builder, the build directory is
somewhere within
~/.var/app/org.gnome.Builder/cache/gnome-builder/projects. As a result,
flake8 does not find its config file, and so it rejects one of our
Python source files for having an 82-character line, despite the .flake8
file specifying a maximum line length of 88.

Tell Meson to run flake8 from the source directory.

The documentation for meson.current_source_directory() claims:
<https://mesonbuild.com/Reference-manual_builtin_meson.html#mesoncurrent_source_dir>
claims:

> Note: You do not need to use this function!

In this case this is actually not true. If you specify the bare path:

    workdir: '.'

Meson rejects it:

> test keyword argument "workdir" must be an absolute path

If you try using files() as the documentation suggests:

    workdir: files('.')

Meson also rejects this:

> File . does not exist.