Tell Meson which tests produce TAP output
GLib, and the Python tests with the taptestrunner.py we copy-pasta, emit
output that conforms to Test Anything Protocol
<https://testanything.org/>, allowing the runner to know the number of
subtests and which of them passed.

With this information, Meson can produce this information to the console
when you run `meson test`:

 1/20 eos-updater:libeos-updater-util / avahi-service-file    OK   0.04s   4 subtests passed
 2/20 eos-updater:libeos-updater-util / ostree-util           OK   0.03s   1 subtests passed

I believe it also includes this information in the JUnit XML it emits
and that our Jenkins consumes.

https://phabricator.endlessm.com/T35247