Remove checks for eos3.9 → eos4 and eos4 → eos5
If a device is running this branch of the updater, it has already passed
these checks; and they are conditional on the source/target branches of
a checkpoint so will not be checked again.

Remove these checks, and all the tests that exercise them. Also remove
all the machinery in the test suite to support overriding these various
fields, which are now unused.

In order to exercise the case where should_follow_checkpoint() returns
FALSE, handle EOS_UPDATER_FORCE_FOLLOW_CHECKPOINT=0 as the opposite of
EOS_UPDATER_FORCE_FOLLOW_CHECKPOINT=1, and keep just enough of the
former tests to handle that.

As we add real checks for the eos5 → eos6 upgrade, these will be tested
by calling this function in isolation, rather than by running the
updater as a whole.

https://phabricator.endlessm.com/T35247