Move checkpoint-checking function to util library
This will allow it to be called from a test without running the entire
daemon.

The OstreeRepo *repo parameter is removed because it can be accessed
with ostree_sysroot_get_repo() if needed.

The OstreeDeployment *booted_deployment parameter is removed partly
because it could be looked up if needed, but more to the point because
it's easier to not have to construct it in the forthcoming test suite!

https://phabricator.endlessm.com/T35247