Record a metrics event when crossing a checkpoint is blocked
This supersedes EOS_UPDATER_METRIC_BRANCH_SELECTED, which has been
ignored on the server since eos4.0. This event contained a boolean
indicating whether the device was being held back from updating, which
was at one point implemented by the updater sending the device vendor &
product to a server, which would instruct it whether to hold back the
update. That functionality was removed long ago, around the time we
implemented USB and LAN updates, so the flag was a stub. What remained
is a singular event recorded at an unspecified frequency.

The mechanism we have now, with ad-hoc checks in eos-updater to prevent
devices crossing a checkpoint when they rely on functionality that has
been removed, is the spiritual successor to that mechanism. But
because being blocked from crossing a checkpoint is not considered an
error, we don't get anything from the existing "updater failed" event
when this happens.

The new event is aggregated on the client side, in the metrics daemon,
at daily and monthly frequency. This will let us count how many devices
are held back from upgrading to EOS 6 due to using nvme-remapped
storage.

The test suite only covers the metrics-disabled path. I have manually
tested the metrics-enabled path and the expected events (daily and
monthly) appear in the client-side metrics daemon's database.

https://phabricator.endlessm.com/T35247
https://phabricator.endlessm.com/T35250