ctl: Update cached State in StateChanged handler
eos-updater emits two signals when its State property changes, in this
order:

- com.endlessm.Updater.StateChanged(x)
- org.freedesktop.DBus.Properties.PropertiesChanged({"State": x, ...})

In eos-updater-ctl, the function that runs a single command works as
follows:

1. Call the method for that command (which returns quickly)
2. Spin the mainloop until StateChanged signals a transition to either
   the target state for the command, or to the Error state
3. Dump all the cached properties of the GDBusProxy to the console
   (unless `--quiet` is passed)
4. Return

At step 3 the PropertiesChanged signal has not been handled (internally
in GDBusProxy, or by us) to update the cached properties.  As a result
you see confusing output like the following:

    $ eos-updater-ctl update
    Changed state to Polling
    Changed state to UpdateAvailable
    ======= Properties =======
                  State: Polling
                    ...: ...

Explicitly update the cached State property when we know it has changed.